Small changes to functions in the root of srcs directory.

+++ srcs/save_history.c
@@ -6,7 +6,7 @@
 /*   By: vkuokka <vkuokka@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/21 18:50:52 by vkuokka           #+#    #+#             */
-/*   Updated: 2020/07/22 15:43:40 by vkuokka          ###   ########.fr       */
+/*   Updated: 2020/07/23 20:50:14 by vkuokka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void		save_history(t_terminal *term)
 {
 	int		fd;
 	t_dlist	*current;
-	int		count;
+	size_t	count;
 
 	fd = open(HISTFILE, O_WRONLY | O_CREAT, 0666);
 	if (fd == -1)
